Crate maidsafe_types

Safe Network Data Types

This library implements the fundimental data types used on the SAFE Network The serialisation mechnism used is cbor` which is an IETF Rfc 7049 for serialising data and is an attempt to upgrade messagepack and ASN.1 On disk serialisation is JSON

Project github page

Reexports

pub use id::{Maid, Mpid, AnMaid, PublicAnMaid, AnMpid, PublicMaid, PublicMpid};

Modules

common

data

helper

id

traits

Structs

ImmutableData
ImmutableData

NameType
NameType struct

Payload
Encoded type serialised and ready to send on wire

StructuredData
StructuredData

Enums

CryptoError

PayloadTypeTag
Types of payload that will be exchange among vaults MaidManager : PublicMaid, PublicAnMaid All : Datatype -- ImmutableData, StructuredData

Traits

Random