Crate maidsafe_types

Safe Network Data Types

This library implements the fundimental data types used on the SAFE Network The serialisation mechnism used is cbor` which is an IETF Rfc 7049 for serialising data and is an attempt to upgrade messagepack and ASN.1 On disk serialisation is JSON

Project github page

Reexports

pub use id::{RevocationIdType, IdType, PublicIdType};

Modules

coin
SafeCoin related details

data
Holds the structs related to data such as ImmutableData/Backup/Sacrificial and StructuredData

helper
Helper provides helper functions for array to vector conversions and vice versa

id
Holds the structs for Id related Types such as Maid, AnMaid, Mpid, etc

Structs

ImmutableData
ImmutableData

ImmutableDataBackup
ImmutableDataBackup

ImmutableDataSacrificial
ImmutableDataSacrificial

MaidTypeTags
TypeTags for Maid type variants

MpidTypeTags
TypeTags for Maid type variants

Payload
Encoded type serialised and ready to send on wire

StructuredData
StructuredData

Enums

CryptoError
Crypto Error types

PayloadTypeTag
Types of payload that will be exchange among vaults MaidManager : PublicMaid, PublicAnMaid All : Datatype -- ImmutableData, ImmutableDataBackup, ImmutableDataSacrificial, StructuredData

Traits

IdTypeTags
Interface to IdTypes

Random
Random trait is used to generate random instances. Used in the test mod

TypeTag
TypeTag trait