Crate maidsafe_types

Placeholder DO NOT USE ! until version 0.1 - all code is a test and useless Types of data functors and messages for MaidSafe secure Autonomous networks. This crate is of no use to anyone as a stand alone crate. It is a module that is specialised, but it is a crate to make version handling and distribution easier.

Reexports

pub use self::name_type::*;

pub use self::an_maid::*;

pub use self::an_mpid::*;

pub use self::maid::*;

pub use self::mpid::*;

pub use self::public_maid::*;

pub use self::public_mpid::*;

pub use self::public_an_maid::*;

pub use self::structured_data::*;

pub use self::immutable_data::*;

Modules

helper

traits